requirements.txt
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
Flask==3.1.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Alembic==1.14.1
Mako==1.3.9
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysqlclient==2.2.8
pip==25.1.1
pycparser==3.0
PyMySQL==1.1.2
python-dotenv==1.2.2
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.6